+++ semialign-indexed/src/Data/Semialign/Indexed.hs
@@ -5,10 +5,10 @@
 -- | Zipping and aligning of indexed functors.
 module Data.Semialign.Indexed (
     SemialignWithIndex (..),
-    SemizipWithIndex (..),
+    ZipWithIndex (..),
     ) where
 
-import Prelude hiding (zip, zipWith)
+import Prelude hiding (zip, zipWith, repeat)
 
 import Control.Lens (FunctorWithIndex (imap))
 
@@ -39,8 +39,8 @@ class (FunctorWithIndex i f, Semialign f) => SemialignWithIndex i f | f -> i whe
     ialignWith :: (i -> These a b -> c) -> f a -> f b -> f c
     ialignWith f a b = imap f (align a b)
 
--- | Indexed version of 'Semizip'
-class (SemialignWithIndex i f, Semizip f) => SemizipWithIndex i f | f -> i where
+-- | Indexed version of 'Zip'
+class (SemialignWithIndex i f, Zip f) => ZipWithIndex i f | f -> i where
     -- | Analogous to 'zipWith', but also provides an index.
     izipWith :: (i -> a -> b -> c) -> f a -> f b -> f c
     izipWith f a b = imap (uncurry . f) (zip a b)
@@ -50,25 +50,25 @@ class (SemialignWithIndex i f, Semizip f) => SemizipWithIndex i f | f -> i where
 -------------------------------------------------------------------------------
 
 instance SemialignWithIndex () Maybe
-instance SemizipWithIndex () Maybe
+instance ZipWithIndex () Maybe
 instance SemialignWithIndex Int []
-instance SemizipWithIndex Int []
+instance ZipWithIndex Int []
 instance SemialignWithIndex Int ZipList
-instance SemizipWithIndex Int ZipList
+instance ZipWithIndex Int ZipList
 
 -------------------------------------------------------------------------------
 -- transformers
 -------------------------------------------------------------------------------
 
 instance SemialignWithIndex () Identity
-instance SemizipWithIndex () Identity
+instance ZipWithIndex () Identity
 
 instance (SemialignWithIndex i f, SemialignWithIndex j g) => SemialignWithIndex (Either i j) (Product f g) where
     ialignWith f (Pair fa ga) (Pair fb gb) = Pair fc gc where
         fc = ialignWith (f . Left) fa fb
         gc = ialignWith (f . Right) ga gb
 
-instance (SemizipWithIndex i f, SemizipWithIndex j g) => SemizipWithIndex (Either i j) (Product f g) where
+instance (ZipWithIndex i f, ZipWithIndex j g) => ZipWithIndex (Either i j) (Product f g) where
     izipWith f (Pair fa ga) (Pair fb gb) = Pair fc gc where
         fc = izipWith (f . Left) fa fb
         gc = izipWith (f . Right) ga gb
@@ -79,7 +79,7 @@ instance (SemialignWithIndex i f, SemialignWithIndex j g) => SemialignWithIndex 
         g i (That gb)     = imap (\j -> f (i, j) . That) gb
         g i (These ga gb) = ialignWith (\j -> f (i, j)) ga gb
 
-instance (SemizipWithIndex i f, SemizipWithIndex j g) => SemizipWithIndex (i, j) (Compose f g) where
+instance (ZipWithIndex i f, ZipWithIndex j g) => ZipWithIndex (i, j) (Compose f g) where
     izipWith f (Compose fga) (Compose fgb) = Compose fgc where
         fgc = izipWith (\i -> izipWith (\j -> f (i, j))) fga fgb
 
@@ -88,12 +88,12 @@ instance (SemizipWithIndex i f, SemizipWithIndex j g) => SemizipWithIndex (i, j)
 -------------------------------------------------------------------------------
 
 instance SemialignWithIndex Int Seq
-instance SemizipWithIndex Int Seq
+instance ZipWithIndex Int Seq
 instance SemialignWithIndex Int IntMap
-instance SemizipWithIndex Int IntMap where
+instance ZipWithIndex Int IntMap where
     izipWith = IntMap.intersectionWithKey
 instance Ord k => SemialignWithIndex k (Map k) where
-instance Ord k => SemizipWithIndex k (Map k) where
+instance Ord k => ZipWithIndex k (Map k) where
     izipWith = Map.intersectionWithKey
 
 -------------------------------------------------------------------------------
@@ -101,7 +101,7 @@ instance Ord k => SemizipWithIndex k (Map k) where
 -------------------------------------------------------------------------------
 
 instance (Eq k, Hashable k) => SemialignWithIndex k (HashMap k) where
-instance (Eq k, Hashable k) => SemizipWithIndex k (HashMap k) where
+instance (Eq k, Hashable k) => ZipWithIndex k (HashMap k) where
     izipWith = HM.intersectionWithKey
 
 -------------------------------------------------------------------------------
@@ -109,5 +109,5 @@ instance (Eq k, Hashable k) => SemizipWithIndex k (HashMap k) where
 -------------------------------------------------------------------------------
 
 instance SemialignWithIndex Int Vector where
-instance SemizipWithIndex Int Vector where
+instance ZipWithIndex Int Vector where
     izipWith = V.izipWith